Fixed the bug in interrupt.c
There was a problem in the way inline assembly was executed, so I changed how the variables are taken from the registers.

I am not really sure if it is a great way, because I read the syntax might be compiler-dependent, but using registers is already pretty system-dependent, so it shouldn't matter.

Also tested more the syscalls code, and it still behaves weirdly, so I made the comment regarding the issue more descriptive.